Add files via upload
currently working on disabling walking past barriers. values are all correct but program keeps crashing because apparently i'm trying to access something outside the map. (map is now a char ** by default for my own sanity).

maybe i have a typo somewhere idk. if u try with map.cub from valid maps it works but crashes as soon as u try to walk. the values printed last are what index is being accessed and it should be fine but it crashes none the less.